requirements.txt
boto3 >= 1.12.6
inquirer >= 2.6.3
python-dateutil >=2.8.1
bs4
lxml